+++ performance_test/src/communication_abstractions/connext_dds_communicator.hpp
@@ -113,10 +113,7 @@ public:
     register_topic();
   }
 
-  void publish(
-    std::int64_t time,
-    std::chrono::duration<double> remaining_time_to_publish =
-    std::chrono::duration<double>{}) override
+  void publish(std::int64_t time) override
   {
     if (m_datawriter == nullptr) {
       DDSPublisher * publisher;
@@ -147,7 +144,7 @@ public:
     }
     m_stats.lock();
     init_msg(m_data, time);
-    m_stats.update_publisher_stats(remaining_time_to_publish);
+    m_stats.update_publisher_stats();
     m_stats.unlock();
     auto retcode = m_typed_datawriter->write(m_data, DDS_HANDLE_NIL);
     if (retcode != DDS_RETCODE_OK) {