+++ performance_test/src/communication_abstractions/iceoryx_communicator.hpp
@@ -41,10 +41,7 @@ public:
   explicit IceoryxCommunicator(DataStats & stats)
   : Communicator(stats) {}
 
-  void publish(
-    std::int64_t time,
-    std::chrono::duration<double> remaining_time_to_publish =
-    std::chrono::duration<double>{}) override
+  void publish(std::int64_t time) override
   {
     if (m_publisher == nullptr) {
       ResourceManager::get().init_iceoryx_runtime();
@@ -61,7 +58,7 @@ public:
         [&](auto & sample) {
           m_stats.lock();
           init_msg(*sample, time);
-          m_stats.update_publisher_stats(remaining_time_to_publish);
+          m_stats.update_publisher_stats();
           m_stats.unlock();
           sample.publish();
         })
@@ -72,7 +69,7 @@ public:
     } else {
       m_stats.lock();
       init_msg(m_data, time);
-      m_stats.update_publisher_stats(remaining_time_to_publish);
+      m_stats.update_publisher_stats();
       m_stats.unlock();
       m_publisher->publishCopyOf(m_data)
       .or_else(